requirements.txt
# Web框架和数据展示
streamlit>=1.28.0
streamlit-pdf-viewer>=0.0.1

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# 地理空间数据处理
geopandas>=0.14.0
shapely>=2.0.0
geopy>=2.4.0

# 地图可视化
folium>=0.15.0
contextily>=1.4.0

# 网络分析
networkx>=3.1

# 空间分析和聚类
libpysal>=4.9.0
spopt>=0.5.0

# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# 深度学习
torch>=2.0.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 图像处理
Pillow>=10.0.0

# 网络请求和数据抓取
overpy>=0.7

# 进度条
tqdm>=4.66.0
